pyproject.toml
[build-system]
requires = ["hatchling>=1.26"]
build-backend = "hatchling.build"

[project]
name = "python-project-template"
version = "0.1.0"
description = "Template em Python para iniciar novos projetos com estrutura organizada, testes, Docker e CI/CD."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Higor Campos" }]

dependencies = [
  "typer>=0.12",
  "python-dotenv>=1.0",
  "pydantic>=2.7",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-cov>=5.0",
  "ruff>=0.6",
  "black>=24.8",
  "mypy>=1.10",
]

[project.scripts]
my-project = "my_project.cli:app"

# Ruff configuration (nova sintaxe)
[tool.ruff]
line-length = 100
target-version = "py311"
exclude = ["docs", "build", "dist"]

[tool.ruff.lint]
extend-select = ["I"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra -q --cov=src --cov-report=term-missing"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
strict = false
warn_unused_ignores = true
ignore_missing_imports = true

[tool.coverage.run]
branch = true
source = ["src/my_project"]

[tool.hatch.build.targets.wheel]
packages = ["src/my_project"]

[tool.hatch.build.targets.sdist]
include = [
  "src/my_project",
  "pyproject.toml",
  "README.md",
  "LICENSE"
]
